+++ bin/image-guard.js
@@ -1,10 +1,15 @@
 #!/usr/bin/env node
 
 import { runImageGuard } from '../src/index.js'
+import { styleStderr } from '../src/utils.js'
 
 try {
   await runImageGuard()
 } catch (err) {
-  console.error('Error running Image Guard:', err.message)
+  if (err.setupFailed) {
+    console.error(styleStderr('red', err.message))
+  } else {
+    console.error('Error running Image Guard:', err)
+  }
   process.exit(1)
 }